backend/requirements.txt
Flask>=2.2
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
cryptography
Pillow
numpy
scikit-image
redis
gunicorn
python-dotenv
pytest
